+++ Reel_n_Roll/src/main/java/com/nichia/reel_n_roll/repository/SeatRepository.java
@@ -1,6 +1,5 @@
 package com.nichia.reel_n_roll.repository;
 
-
 // Tanya Navani - Student ID: 405008840
 
 import com.nichia.reel_n_roll.model.Seat;
@@ -13,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
 
     List<Seat> findBySeatNumberContainingIgnoreCase(String seatNumber);
 
@@ -25,5 +24,5 @@ public interface SeatRepository extends JpaRepository<Seat, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE Seat s SET s.booked = :booked WHERE s.seatId = :seatId")
-    int updateBookedBySeatId(Long seatId, boolean booked);
+    int updateBookedBySeatId(Integer seatId, boolean booked);
 }
